+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 0.4.0
 
-- Experimental: The default scaling logic is contained in ZeroOneScalingStrategy.  A strategy object can be passed instead of timeout to the server middleware.
+- Experimental: The default scaling logic is contained in BinaryScalingStrategy.  A strategy object can be passed instead of timeout to the server middleware.
 
 ## 0.3.0